refactor: rename user_repository.py to user.py
